Always allow switches with no cases
Exhaustiveness checking is conversative: there are many cases that the compiler won't catch.

This may not be an unreasonable thing to write:

```swift
enum E<V> {
  case one(V)
  case two(V, V)
}

func f(e: E<Never>) {
  switch e {}
}
```

(`f` may be part of a protocol and `Never` might make sense for some type that conforms to it.)

Remove check that all `switch`s that aren't known to be uninhabited must have at least one `case`. If cases are missing, this will be caught as part of exhaustiveness checking when the `switch` is well-formed.